create separate volume control for Metronome
